requirements.txt
frappe
pytesseract
Pillow
setuptools
wheel

# System dependencies
tesseract-ocr